Add StaticDeserialization class
