+++ geode-membership/src/main/java/org/apache/geode/distributed/internal/membership/gms/GMSMembershipView.java
@@ -38,6 +38,7 @@ import org.apache.geode.internal.serialization.DataSerializableFixedID;
 import org.apache.geode.internal.serialization.DeserializationContext;
 import org.apache.geode.internal.serialization.KnownVersion;
 import org.apache.geode.internal.serialization.SerializationContext;
+import org.apache.geode.internal.serialization.StaticDeserialization;
 import org.apache.geode.internal.serialization.StaticSerialization;
 
 /**
@@ -620,8 +621,8 @@ public class GMSMembershipView<ID extends MemberIdentifier> implements DataSeria
     this.hashedMembers = new HashSet<>(members);
     shutdownMembers = GMSUtil.readHashSetOfMemberIDs(in, context);
     crashedMembers = GMSUtil.readHashSetOfMemberIDs(in, context);
-    failureDetectionPorts = StaticSerialization.readIntArray(in);
-    Map<ID, Object> pubkeys = StaticSerialization.readHashMap(in, context);
+    failureDetectionPorts = StaticDeserialization.readIntArray(in);
+    Map<ID, Object> pubkeys = StaticDeserialization.readHashMap(in, context);
     if (pubkeys != null) {
       publicKeys.putAll(pubkeys);
     }
