+++ geode-serialization/src/main/java/org/apache/geode/internal/serialization/StaticSerialization.java
@@ -18,15 +18,13 @@ import java.io.DataInput;
 import java.io.DataOutput;
 import java.io.IOException;
 import java.net.InetAddress;
-import java.net.UnknownHostException;
-import java.util.HashMap;
 import java.util.Map;
 
 import org.apache.geode.annotations.internal.MakeNotStatic;
 
 /**
- * StaticSerialization provides a collection of serialization/deserialization methods that
- * can be used in your toData/fromData methods in a DataSerializableFixedID implementation.
+ * StaticSerialization provides a collection of serialization methods that
+ * can be used in your toData methods in a DataSerializableFixedID implementation.
  */
 public class StaticSerialization {
   // array is null
@@ -107,29 +105,6 @@ public class StaticSerialization {
     }
   }
 
-  public static int readArrayLength(DataInput in) throws IOException {
-    byte code = in.readByte();
-    if (code == NULL_ARRAY) {
-      return -1;
-    } else {
-      int result = ubyteToInt(code);
-      if (result > MAX_BYTE_ARRAY_LEN) {
-        if (code == SHORT_ARRAY_LEN) {
-          return in.readUnsignedShort();
-        } else if (code == INT_ARRAY_LEN) {
-          return in.readInt();
-        } else {
-          throw new IllegalStateException("unexpected array length code=" + code);
-        }
-      }
-      return result;
-    }
-  }
-
-
-  private static int ubyteToInt(byte ub) {
-    return ub & 0xFF;
-  }
 
   public static void writeString(String value, DataOutput out) throws IOException {
 
@@ -186,96 +161,6 @@ public class StaticSerialization {
     }
   }
 
-  public static String readString(DataInput in, byte header) throws IOException {
-    return readString(in, DscodeHelper.toDSCODE(header));
-  }
-
-  private static String readString(DataInput in, DSCODE dscode) throws IOException {
-    switch (dscode) {
-      case STRING_BYTES:
-        return readStringBytesFromDataInput(in, in.readUnsignedShort());
-      case STRING:
-        return readStringUTFFromDataInput(in);
-      case NULL_STRING: {
-        return null;
-      }
-      case HUGE_STRING_BYTES:
-        return readStringBytesFromDataInput(in, in.readInt());
-      case HUGE_STRING:
-        return readHugeStringFromDataInput(in);
-      default:
-        throw new IOException("Unknown String header " + dscode);
-    }
-  }
-
-  private static String readHugeStringFromDataInput(DataInput in) throws IOException {
-    int len = in.readInt();
-    char[] buf = new char[len];
-    for (int i = 0; i < len; i++) {
-      buf[i] = in.readChar();
-    }
-    return new String(buf);
-  }
-
-  public static String[] readStringArray(DataInput in) throws IOException {
-
-    int length = readArrayLength(in);
-    if (length == -1) {
-      return null;
-    } else {
-      String[] array = new String[length];
-      for (int i = 0; i < length; i++) {
-        array[i] = readString(in);
-      }
-      return array;
-    }
-  }
-
-  private static String readStringUTFFromDataInput(DataInput in) throws IOException {
-    return in.readUTF();
-  }
-
-
-
-  private static String readStringBytesFromDataInput(DataInput dataInput, int len)
-      throws IOException {
-    if (len == 0) {
-      return "";
-    }
-    byte[] buf = getThreadLocalByteArray(len);
-    dataInput.readFully(buf, 0, len);
-    @SuppressWarnings("deprecation") // intentionally using deprecated constructor
-    final String string = new String(buf, 0, 0, len);
-    return string;
-  }
-
-  /**
-   * Reads an instance of <code>String</code> from a <code>DataInput</code>. The return value may be
-   * <code>null</code>.
-   *
-   * @throws IOException A problem occurs while reading from <code>in</code>
-   *
-   * @see #writeString
-   */
-  public static String readString(DataInput in) throws IOException {
-    return readString(in, in.readByte());
-  }
-
-  public static InetAddress readInetAddress(DataInput in) throws IOException {
-
-    byte[] address = readByteArray(in);
-    if (address == null) {
-      return null;
-    }
-
-    try {
-      // note: this does not throw UnknownHostException at this time
-      return InetAddress.getByAddress(address);
-    } catch (UnknownHostException ex) {
-      throw new IOException("While reading an InetAddress", ex);
-    }
-
-  }
 
   public static void writeStringArray(String[] array, DataOutput out) throws IOException {
     int length;
@@ -296,49 +181,6 @@ public class StaticSerialization {
     out.writeInt(value);
   }
 
-  public static <K, V> HashMap<K, V> readHashMap(DataInput in, DeserializationContext context)
-      throws IOException, ClassNotFoundException {
-
-    int size = readArrayLength(in);
-    if (size == -1) {
-      return null;
-    } else {
-      HashMap<K, V> map = new HashMap<>(size);
-      for (int i = 0; i < size; i++) {
-        K key = context.getDeserializer().readObject(in);
-        V value = context.getDeserializer().readObject(in);
-        map.put(key, value);
-      }
-
-      return map;
-    }
-  }
-
-
-  public static int[] readIntArray(DataInput in) throws IOException {
-    int length = readArrayLength(in);
-    if (length == -1) {
-      return null;
-    } else {
-      int[] array = new int[length];
-      for (int i = 0; i < length; i++) {
-        array[i] = in.readInt();
-      }
-      return array;
-    }
-  }
-
-  public static byte[] readByteArray(DataInput in) throws IOException {
-
-    int length = readArrayLength(in);
-    if (length == -1) {
-      return null;
-    } else {
-      byte[] array = new byte[length];
-      in.readFully(array, 0, length);
-      return array;
-    }
-  }
 
   public static void writeIntArray(int[] array, DataOutput out) throws IOException {
     int length;
@@ -375,18 +217,6 @@ public class StaticSerialization {
   }
 
 
-
-  /**
-   * Returns a byte array for use by the calling thread.
-   * The returned byte array may be longer than minimumLength.
-   * The byte array belongs to the calling thread but callers must
-   * be careful to not call other methods that may also use this
-   * byte array.
-   */
-  public static byte[] getThreadLocalByteArray(int minimumLength) {
-    return threadLocalByteArrayCache.get(minimumLength);
-  }
-
   public static void writeClass(final Class<?> c, final DataOutput out) throws IOException {
     if (c == null || c.isPrimitive()) {
       writePrimitiveClass(c, out);
@@ -402,17 +232,6 @@ public class StaticSerialization {
     }
   }
 
-  public static Class<?> readClass(DataInput in) throws IOException, ClassNotFoundException {
-    byte typeCode = in.readByte();
-    if (typeCode == DSCODE.CLASS.toByte()) {
-      String className = readString(in);
-      className = processIncomingClassName(className);
-      return Class.forName(className);
-    } else {
-      return StaticSerialization.decodePrimitiveClass(typeCode);
-    }
-  }
-
   /**
    * Map from new package to old package.
    *
@@ -424,21 +243,6 @@ public class StaticSerialization {
     if (name.startsWith(POST_GEODE_100_TCPSERVER_PACKAGE)) {
       return PRE_GEODE_100_TCPSERVER_PACKAGE
           + name.substring(POST_GEODE_100_TCPSERVER_PACKAGE.length());
-    }
-    return name;
-  }
-
-  /**
-   * Map from old package to new package.
-   *
-   * @return the same name String (identity) if the package name does not need to change
-   */
-  public static String processIncomingClassName(String name) {
-    // TCPServer classes are used before a cache exists and support for old clients has been
-    // initialized
-    if (name.startsWith(StaticSerialization.PRE_GEODE_100_TCPSERVER_PACKAGE)) {
-      return StaticSerialization.POST_GEODE_100_TCPSERVER_PACKAGE
-          + name.substring(StaticSerialization.PRE_GEODE_100_TCPSERVER_PACKAGE.length());
     }
     return name;
   }
@@ -471,35 +275,6 @@ public class StaticSerialization {
       throw new IllegalArgumentException(
           String.format("unknown primitive type: %s",
               c.getName()));
-    }
-  }
-
-  public static Class<?> decodePrimitiveClass(byte typeCode) throws IOException {
-    DSCODE dscode = DscodeHelper.toDSCODE(typeCode);
-    switch (dscode) {
-      case BOOLEAN_TYPE:
-        return Boolean.TYPE;
-      case CHARACTER_TYPE:
-        return Character.TYPE;
-      case BYTE_TYPE:
-        return Byte.TYPE;
-      case SHORT_TYPE:
-        return Short.TYPE;
-      case INTEGER_TYPE:
-        return Integer.TYPE;
-      case LONG_TYPE:
-        return Long.TYPE;
-      case FLOAT_TYPE:
-        return Float.TYPE;
-      case DOUBLE_TYPE:
-        return Double.TYPE;
-      case VOID_TYPE:
-        return Void.TYPE;
-      case NULL:
-        return null;
-      default:
-        throw new IllegalArgumentException(
-            String.format("unexpected typeCode: %s", typeCode));
     }
   }
